backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3

# Message Queue
stomp.py==8.1.2
paho-mqtt==1.6.1

# Authentication & Authorization
python-jose[cryptography]==3.3.0
cryptography==43.0.1
pyjwt[crypto]==2.9.0
python-multipart==0.0.12
msal==1.31.0

# HTTP Client
httpx==0.27.2
oauthlib==3.2.2

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
langchain-openai==0.2.5

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
black==24.10.0
ruff==0.7.0
mypy==1.13.0
